[feat]: 공통 상수 저장하는 Const 파일 추가
