+++ Todoary-iOS/Todoary/CommonSource/Const.swift
@@ -1,0 +1,23 @@
+//
+//  Const.swift
+//  Todoary
+//
+//  Created by 박지윤 on 2022/09/22.
+//
+
+import Foundation
+
+class Const{
+    
+    class Device{
+        static let DEVICE_HEIGHT = UIScreen.main.bounds.size.height
+        static let DEVICE_WIDTH = UIScreen.main.bounds.size.width
+    }
+    
+    class UserDefaults{
+        static let appleRefreshToken = "APPLE_REFRESH_TOKEN"
+        static let appleIdentifier = "APPLE_IDENTIFIER"
+        static let email = "APPLE_EMAIL"
+        static let userName = "APPLE_USERNAME"
+    }
+}
