+++ Todoary-iOS/Todoary/Home/Summary/SummaryBottomSheetViewController+Layout.swift
@@ -125,10 +125,8 @@ extension SummaryBottomViewController: UIViewControllerTransitioningDelegate{
         
         let controller: UISheetPresentationController = .init(presentedViewController: presented, presenting: presenting)
         
-        let height = UIScreen.main.bounds.size.height
-        
-        let detent1: UISheetPresentationController.Detent = ._detent(withIdentifier: "Test1", constant: height - 519 - 41)
-        let detent2: UISheetPresentationController.Detent = ._detent(withIdentifier: "Test2", constant: height - 95 - 41 ) //790 - 41
+        let detent1: UISheetPresentationController.Detent = ._detent(withIdentifier: "Test1", constant: Const.Device.DEVICE_HEIGHT - (519 + 41))
+        let detent2: UISheetPresentationController.Detent = ._detent(withIdentifier: "Test2", constant: Const.Device.DEVICE_HEIGHT - (95 + 41) ) //790 - 41
         
         let detentIdentifier :UISheetPresentationController.Detent.Identifier = UISheetPresentationController.Detent.Identifier(rawValue: "Test2")
